requirements.txt
python-telegram-bot>=20.7
pyairtable>=2.2.0
redis>=5.0.0
apscheduler>=3.10.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0
